test-requirements.txt
pytest
pytest-cov
vcrpy
requests_mock
